Revert "unify f32"
This reverts commit 89f7f361c4e5ae68bc55e8b28f2359eab0bd49fc.

+++ src/main.rs
@@ -89,7 +89,7 @@ mod domain {
 
     pub trait PaymentClassification: DynClone + Debug {
         fn as_any_mut(&mut self) -> &mut dyn Any;
-        fn calculate_pay(&self, pc: &PayCheck) -> f32;
+        fn calculate_pay(&self, pc: &PayCheck) -> f64;
     }
     dyn_clone::clone_trait_object!(PaymentClassification);
 
@@ -108,7 +108,7 @@ mod domain {
     pub trait Affiliation: DynClone + Debug {
         fn as_any(&self) -> &dyn Any;
         fn as_any_mut(&mut self) -> &mut dyn Any;
-        fn calculate_deductions(&self, _pc: &PayCheck) -> f32 {
+        fn calculate_deductions(&self, _pc: &PayCheck) -> f64 {
             0.0
         }
     }
@@ -118,9 +118,9 @@ mod domain {
     pub struct PayCheck {
         period: RangeInclusive<NaiveDate>,
 
-        gross_pay: f32,
-        deductions: f32,
-        net_pay: f32,
+        gross_pay: f64,
+        deductions: f64,
+        net_pay: f64,
     }
     impl PayCheck {
         pub fn new(period: RangeInclusive<NaiveDate>) -> Self {
@@ -134,13 +134,13 @@ mod domain {
         pub fn get_pay_period(&self) -> RangeInclusive<NaiveDate> {
             self.period.clone()
         }
-        pub fn set_gross_pay(&mut self, gross_pay: f32) {
+        pub fn set_gross_pay(&mut self, gross_pay: f64) {
             self.gross_pay = gross_pay;
         }
-        pub fn set_deductions(&mut self, deductions: f32) {
+        pub fn set_deductions(&mut self, deductions: f64) {
             self.deductions = deductions;
         }
-        pub fn set_net_pay(&mut self, net_pay: f32) {
+        pub fn set_net_pay(&mut self, net_pay: f64) {
             self.net_pay = net_pay;
         }
     }
@@ -214,31 +214,31 @@ mod classification {
 
     #[derive(Debug, Clone, PartialEq)]
     pub struct SalariedClassification {
-        salary: f32,
+        salary: f64,
     }
     impl PaymentClassification for SalariedClassification {
         fn as_any_mut(&mut self) -> &mut dyn Any {
             self
         }
-        fn calculate_pay(&self, _pc: &PayCheck) -> f32 {
+        fn calculate_pay(&self, _pc: &PayCheck) -> f64 {
             self.salary
         }
     }
     impl SalariedClassification {
-        pub fn new(salary: f32) -> Self {
+        pub fn new(salary: f64) -> Self {
             Self { salary }
         }
     }
     #[derive(Debug, Clone, PartialEq)]
     pub struct HourlyClassification {
-        hourly_rate: f32,
+        hourly_rate: f64,
         timecards: Vec<TimeCard>,
     }
     impl PaymentClassification for HourlyClassification {
         fn as_any_mut(&mut self) -> &mut dyn Any {
             self
         }
-        fn calculate_pay(&self, pc: &PayCheck) -> f32 {
+        fn calculate_pay(&self, pc: &PayCheck) -> f64 {
             let pay_period = pc.get_pay_period();
             let mut total_pay = 0.0;
             for tc in self.timecards.iter() {
@@ -250,7 +250,7 @@ mod classification {
         }
     }
     impl HourlyClassification {
-        pub fn new(hourly_rate: f32) -> Self {
+        pub fn new(hourly_rate: f64) -> Self {
             Self {
                 hourly_rate,
                 timecards: vec![],
@@ -259,7 +259,7 @@ mod classification {
         pub fn add_timecard(&mut self, tc: TimeCard) {
             self.timecards.push(tc);
         }
-        pub fn calculate_pay_for_timecard(&self, tc: &TimeCard) -> f32 {
+        pub fn calculate_pay_for_timecard(&self, tc: &TimeCard) -> f64 {
             let hours = tc.get_hours();
             let overtime = (hours - 8.0).max(0.0);
             let straight_time = hours - overtime;
@@ -268,15 +268,15 @@ mod classification {
     }
     #[derive(Debug, Clone, PartialEq)]
     pub struct CommissionedClassification {
-        salary: f32,
-        commission_rate: f32,
+        salary: f64,
+        commission_rate: f64,
         sales_receipts: Vec<SalesReceipt>,
     }
     impl PaymentClassification for CommissionedClassification {
         fn as_any_mut(&mut self) -> &mut dyn Any {
             self
         }
-        fn calculate_pay(&self, pc: &PayCheck) -> f32 {
+        fn calculate_pay(&self, pc: &PayCheck) -> f64 {
             let mut total_pay = 0.0;
             let pay_period = pc.get_pay_period();
             for sr in self.sales_receipts.iter() {
@@ -288,7 +288,7 @@ mod classification {
         }
     }
     impl CommissionedClassification {
-        pub fn new(salary: f32, commission_rate: f32) -> Self {
+        pub fn new(salary: f64, commission_rate: f64) -> Self {
             Self {
                 salary,
                 commission_rate,
@@ -298,7 +298,7 @@ mod classification {
         pub fn add_sales_receipt(&mut self, sr: SalesReceipt) {
             self.sales_receipts.push(sr);
         }
-        pub fn calculate_pay_for_sales_receipt(&self, sr: &SalesReceipt) -> f32 {
+        pub fn calculate_pay_for_sales_receipt(&self, sr: &SalesReceipt) -> f64 {
             self.commission_rate * sr.get_amount()
         }
     }
@@ -306,16 +306,16 @@ mod classification {
     #[derive(Debug, Clone, PartialEq)]
     pub struct TimeCard {
         date: NaiveDate,
-        hours: f32,
+        hours: f64,
     }
     impl TimeCard {
-        pub fn new(date: NaiveDate, hours: f32) -> Self {
+        pub fn new(date: NaiveDate, hours: f64) -> Self {
             Self { date, hours }
         }
         pub fn get_date(&self) -> NaiveDate {
             self.date
         }
-        pub fn get_hours(&self) -> f32 {
+        pub fn get_hours(&self) -> f64 {
             self.hours
         }
     }
@@ -323,16 +323,16 @@ mod classification {
     #[derive(Debug, Clone, PartialEq)]
     pub struct SalesReceipt {
         date: NaiveDate,
-        amount: f32,
+        amount: f64,
     }
     impl SalesReceipt {
-        pub fn new(date: NaiveDate, amount: f32) -> Self {
+        pub fn new(date: NaiveDate, amount: f64) -> Self {
             Self { date, amount }
         }
         pub fn get_date(&self) -> NaiveDate {
             self.date
         }
-        pub fn get_amount(&self) -> f32 {
+        pub fn get_amount(&self) -> f64 {
             self.amount
         }
     }
@@ -435,12 +435,12 @@ mod affiliation {
     #[derive(Debug, Clone, PartialEq)]
     pub struct UnionAffiliation {
         member_id: MemberId,
-        dues: f32,
+        dues: f64,
 
         service_charges: Vec<ServiceCharge>,
     }
     impl UnionAffiliation {
-        pub fn new(member_id: MemberId, dues: f32) -> Self {
+        pub fn new(member_id: MemberId, dues: f64) -> Self {
             Self {
                 member_id,
                 dues,
@@ -450,7 +450,7 @@ mod affiliation {
         pub fn get_member_id(&self) -> MemberId {
             self.member_id
         }
-        pub fn get_dues(&self) -> f32 {
+        pub fn get_dues(&self) -> f64 {
             self.dues
         }
         pub fn add_service_charge(&mut self, sc: ServiceCharge) {
@@ -464,7 +464,7 @@ mod affiliation {
         fn as_any_mut(&mut self) -> &mut dyn Any {
             self
         }
-        fn calculate_deductions(&self, pc: &PayCheck) -> f32 {
+        fn calculate_deductions(&self, pc: &PayCheck) -> f64 {
             let mut total_deductions = 0.0;
             let pay_period = pc.get_pay_period();
             for d in pc.get_pay_period().start().iter_days() {
@@ -497,16 +497,16 @@ mod affiliation {
     #[derive(Debug, Clone, PartialEq)]
     pub struct ServiceCharge {
         date: NaiveDate,
-        amount: f32,
+        amount: f64,
     }
     impl ServiceCharge {
-        pub fn new(date: NaiveDate, amount: f32) -> Self {
+        pub fn new(date: NaiveDate, amount: f64) -> Self {
             Self { date, amount }
         }
         pub fn get_date(&self) -> NaiveDate {
             self.date
         }
-        pub fn get_amount(&self) -> f32 {
+        pub fn get_amount(&self) -> f64 {
             self.amount
         }
     }
@@ -721,7 +721,7 @@ mod general_tx {
         fn get_emp_id(&self) -> EmployeeId;
         fn get_name(&self) -> &str;
         fn get_address(&self) -> &str;
-        fn get_salary(&self) -> f32;
+        fn get_salary(&self) -> f64;
     }
     pub trait AddSalaryEmployeeTransaction<Ctx>:
         AddEmployeeTransaction<Ctx> + SalaryEmployee
@@ -795,7 +795,7 @@ mod general_tx {
         fn get_emp_id(&self) -> EmployeeId;
         fn get_name(&self) -> &str;
         fn get_address(&self) -> &str;
-        fn get_hourly_rate(&self) -> f32;
+        fn get_hourly_rate(&self) -> f64;
     }
     pub trait AddHourlyEmployeeTransaction<Ctx>:
         AddEmployeeTransaction<Ctx> + HourlyEmployee
@@ -869,8 +869,8 @@ mod general_tx {
         fn get_emp_id(&self) -> EmployeeId;
         fn get_name(&self) -> &str;
         fn get_address(&self) -> &str;
-        fn get_salary(&self) -> f32;
-        fn get_commission_rate(&self) -> f32;
+        fn get_salary(&self) -> f64;
+        fn get_commission_rate(&self) -> f64;
     }
     pub trait AddCommissionedEmployeeTransaction<Ctx>:
         AddEmployeeTransaction<Ctx> + CommissionedEmployee
@@ -999,7 +999,7 @@ mod general_tx {
     pub trait TimeCardEmployee {
         fn get_emp_id(&self) -> EmployeeId;
         fn get_date(&self) -> NaiveDate;
-        fn get_hours(&self) -> f32;
+        fn get_hours(&self) -> f64;
     }
     pub trait TimeCardTransaction<Ctx>: HaveEmployeeDao<Ctx> + TimeCardEmployee {
         fn execute<'a>(&'a self) -> impl tx_rs::Tx<Ctx, Item = (), Err = EmployeeUsecaseError> {
@@ -1068,7 +1068,7 @@ mod general_tx {
     pub trait SalesReceiptEmployee {
         fn get_emp_id(&self) -> EmployeeId;
         fn get_date(&self) -> NaiveDate;
-        fn get_amount(&self) -> f32;
+        fn get_amount(&self) -> f64;
     }
     pub trait SalesReceiptTransaction<Ctx>: HaveEmployeeDao<Ctx> + SalesReceiptEmployee {
         fn execute<'a>(&'a self) -> impl tx_rs::Tx<Ctx, Item = (), Err = EmployeeUsecaseError> {
@@ -1340,7 +1340,7 @@ mod classification_tx {
 
     pub trait SalaryChangeableEmployee {
         fn get_emp_id(&self) -> EmployeeId;
-        fn get_salary(&self) -> f32;
+        fn get_salary(&self) -> f64;
     }
 
     pub trait ChangeSalariedTransaction<Ctx>:
@@ -1403,7 +1403,7 @@ mod classification_tx {
 
     pub trait HourlyChangeableEmployee {
         fn get_emp_id(&self) -> EmployeeId;
-        fn get_hourly_rate(&self) -> f32;
+        fn get_hourly_rate(&self) -> f64;
     }
     pub trait ChangeHourlyTransaction<Ctx>:
         ChangeClassificationTransaction<Ctx> + HourlyChangeableEmployee
@@ -1465,8 +1465,8 @@ mod classification_tx {
 
     pub trait CommissionedChangeableEmployee {
         fn get_emp_id(&self) -> EmployeeId;
-        fn get_salary(&self) -> f32;
-        fn get_commission_rate(&self) -> f32;
+        fn get_salary(&self) -> f64;
+        fn get_commission_rate(&self) -> f64;
     }
     pub trait ChangeCommissionedTransaction<Ctx>:
         ChangeClassificationTransaction<Ctx> + CommissionedChangeableEmployee
@@ -1741,7 +1741,7 @@ mod affiliation_tx {
     pub trait ServiceChargeableMember {
         fn get_member_id(&self) -> MemberId;
         fn get_date(&self) -> NaiveDate;
-        fn get_amount(&self) -> f32;
+        fn get_amount(&self) -> f64;
     }
     pub trait ServiceChargeTransaction<Ctx>:
         HaveEmployeeDao<Ctx> + ServiceChargeableMember
@@ -1821,7 +1821,7 @@ mod affiliation_tx {
     pub trait UnionChangeableEmployee {
         fn get_emp_id(&self) -> EmployeeId;
         fn get_member_id(&self) -> MemberId;
-        fn get_dues(&self) -> f32;
+        fn get_dues(&self) -> f64;
     }
     pub trait ChangeUnionMemberTransaction<Ctx>:
         ChangeAffiliationTransaction<Ctx> + UnionChangeableEmployee
@@ -2066,7 +2066,7 @@ struct AddSalariedEmployeeTransactionImpl {
     emp_id: EmployeeId,
     name: String,
     address: String,
-    salary: f32,
+    salary: f64,
 }
 impl HaveEmployeeDao<()> for AddSalariedEmployeeTransactionImpl {
     fn dao(&self) -> Box<&impl EmployeeDao<()>> {
@@ -2083,7 +2083,7 @@ impl SalaryEmployee for AddSalariedEmployeeTransactionImpl {
     fn get_address(&self) -> &str {
         &self.address
     }
-    fn get_salary(&self) -> f32 {
+    fn get_salary(&self) -> f64 {
         self.salary
     }
 }
@@ -2094,7 +2094,7 @@ struct AddHourlyEmployeeTransactionImpl {
     emp_id: EmployeeId,
     name: String,
     address: String,
-    hourly_rate: f32,
+    hourly_rate: f64,
 }
 impl HaveEmployeeDao<()> for AddHourlyEmployeeTransactionImpl {
     fn dao(&self) -> Box<&impl EmployeeDao<()>> {
@@ -2111,7 +2111,7 @@ impl HourlyEmployee for AddHourlyEmployeeTransactionImpl {
     fn get_address(&self) -> &str {
         &self.address
     }
-    fn get_hourly_rate(&self) -> f32 {
+    fn get_hourly_rate(&self) -> f64 {
         self.hourly_rate
     }
 }
@@ -2122,8 +2122,8 @@ struct AddCommissionedEmployeeTransactionImpl {
     emp_id: EmployeeId,
     name: String,
     address: String,
-    salary: f32,
-    commission_rate: f32,
+    salary: f64,
+    commission_rate: f64,
 }
 impl HaveEmployeeDao<()> for AddCommissionedEmployeeTransactionImpl {
     fn dao(&self) -> Box<&impl EmployeeDao<()>> {
@@ -2140,10 +2140,10 @@ impl CommissionedEmployee for AddCommissionedEmployeeTransactionImpl {
     fn get_address(&self) -> &str {
         &self.address
     }
-    fn get_salary(&self) -> f32 {
+    fn get_salary(&self) -> f64 {
         self.salary
     }
-    fn get_commission_rate(&self) -> f32 {
+    fn get_commission_rate(&self) -> f64 {
         self.commission_rate
     }
 }
@@ -2169,7 +2169,7 @@ struct TimeCardTransactionImpl {
 
     emp_id: EmployeeId,
     date: NaiveDate,
-    hours: f32,
+    hours: f64,
 }
 impl HaveEmployeeDao<()> for TimeCardTransactionImpl {
     fn dao(&self) -> Box<&impl EmployeeDao<()>> {
@@ -2183,7 +2183,7 @@ impl TimeCardEmployee for TimeCardTransactionImpl {
     fn get_date(&self) -> NaiveDate {
         self.date
     }
-    fn get_hours(&self) -> f32 {
+    fn get_hours(&self) -> f64 {
         self.hours
     }
 }
@@ -2193,7 +2193,7 @@ struct SalesReceiptTransactionImpl {
 
     emp_id: EmployeeId,
     date: NaiveDate,
-    amount: f32,
+    amount: f64,
 }
 impl HaveEmployeeDao<()> for SalesReceiptTransactionImpl {
     fn dao(&self) -> Box<&impl EmployeeDao<()>> {
@@ -2207,7 +2207,7 @@ impl SalesReceiptEmployee for SalesReceiptTransactionImpl {
     fn get_date(&self) -> NaiveDate {
         self.date
     }
-    fn get_amount(&self) -> f32 {
+    fn get_amount(&self) -> f64 {
         self.amount
     }
 }
@@ -2217,7 +2217,7 @@ struct ServiceChargeTransactionImpl {
 
     member_id: MemberId,
     date: NaiveDate,
-    amount: f32,
+    amount: f64,
 }
 impl HaveEmployeeDao<()> for ServiceChargeTransactionImpl {
     fn dao(&self) -> Box<&impl EmployeeDao<()>> {
@@ -2231,7 +2231,7 @@ impl ServiceChargeableMember for ServiceChargeTransactionImpl {
     fn get_date(&self) -> NaiveDate {
         self.date
     }
-    fn get_amount(&self) -> f32 {
+    fn get_amount(&self) -> f64 {
         self.amount
     }
 }
@@ -2280,7 +2280,7 @@ struct ChangeSalaryTransactionImpl {
     db: MockDb,
 
     emp_id: EmployeeId,
-    salary: f32,
+    salary: f64,
 }
 impl HaveEmployeeDao<()> for ChangeSalaryTransactionImpl {
     fn dao(&self) -> Box<&impl EmployeeDao<()>> {
@@ -2291,7 +2291,7 @@ impl SalaryChangeableEmployee for ChangeSalaryTransactionImpl {
     fn get_emp_id(&self) -> EmployeeId {
         self.emp_id
     }
-    fn get_salary(&self) -> f32 {
+    fn get_salary(&self) -> f64 {
         self.salary
     }
 }
@@ -2300,7 +2300,7 @@ struct ChangeHourlyTransactionImpl {
     db: MockDb,
 
     emp_id: EmployeeId,
-    hourly_rate: f32,
+    hourly_rate: f64,
 }
 impl HaveEmployeeDao<()> for ChangeHourlyTransactionImpl {
     fn dao(&self) -> Box<&impl EmployeeDao<()>> {
@@ -2311,7 +2311,7 @@ impl HourlyChangeableEmployee for ChangeHourlyTransactionImpl {
     fn get_emp_id(&self) -> EmployeeId {
         self.emp_id
     }
-    fn get_hourly_rate(&self) -> f32 {
+    fn get_hourly_rate(&self) -> f64 {
         self.hourly_rate
     }
 }
@@ -2320,8 +2320,8 @@ struct ChangeCommissionedTransactionImpl {
     db: MockDb,
 
     emp_id: EmployeeId,
-    salary: f32,
-    commission_rate: f32,
+    salary: f64,
+    commission_rate: f64,
 }
 impl HaveEmployeeDao<()> for ChangeCommissionedTransactionImpl {
     fn dao(&self) -> Box<&impl EmployeeDao<()>> {
@@ -2332,10 +2332,10 @@ impl CommissionedChangeableEmployee for ChangeCommissionedTransactionImpl {
     fn get_emp_id(&self) -> EmployeeId {
         self.emp_id
     }
-    fn get_salary(&self) -> f32 {
+    fn get_salary(&self) -> f64 {
         self.salary
     }
-    fn get_commission_rate(&self) -> f32 {
+    fn get_commission_rate(&self) -> f64 {
         self.commission_rate
     }
 }
@@ -2405,7 +2405,7 @@ struct ChangeUnionMemberTransactionImpl {
 
     emp_id: EmployeeId,
     member_id: EmployeeId,
-    dues: f32,
+    dues: f64,
 }
 impl HaveEmployeeDao<()> for ChangeUnionMemberTransactionImpl {
     fn dao(&self) -> Box<&impl EmployeeDao<()>> {
@@ -2419,7 +2419,7 @@ impl UnionChangeableEmployee for ChangeUnionMemberTransactionImpl {
     fn get_member_id(&self) -> EmployeeId {
         self.member_id
     }
-    fn get_dues(&self) -> f32 {
+    fn get_dues(&self) -> f64 {
         self.dues
     }
 }
